walk the commit history by chronological order
This is the same that 'git log' does by default, and it's what we want.

Before, we were doing a post-order walk, stopping if the "main" chain
had a timestamp that was too old. That seemed to work okay for simple
scenarios, but for complex ones involving merging old code, it broke
down easily.

One such example was that we would have sporadic CI failures, if the
machine was slow to run the 'git commit' commands. This was reproducible
if one added sleeps in between each of them.

Instead, walk the commits in chronological order by commit date. This is
what 'git log' does, and we can safely stop the walking when we go too
far back in time.

As a bonus, we can stop keeping track of the "main chain" when comparing
timestamps.