add support for Gerrit-like Change-Ids
This can be useful if one updates a commit's metadata while merging,
such as amending its title or commit message.

Since Change-Id lines should be unique per repository, and since the
logic is unchanged for commits that don't have a Change-Id, there should
be no regressions. The advantage is that, if a developer uses
Change-Ids, the tool should be more precise now.

Thanks to Paul Jolly for the suggestion.